refactor: Update import/require match logic
1. Using importSource array instead of importName
This helps if project mixes usages between graphql-tag and apollo v3
This mix is uncouraged but it could happen in big projects with unmantained modules.

1. Match require instead of only es6 import
2. Match gql inside of any `import {} from X;`

Downsides: Some cases are really hard to track down:

```
const apolloClient = require('@apollo/client');
// or
import apolloClient from '@apollo/client';

const { gql } = apolloClient;

const foo = gql`...`;
```

Note: add a disclaimer that it is a known issue and this kind of syntax is not supported.